feat(posts): structure posts with components
