+++ src/components/post/description.tsx
@@ -1,0 +1,11 @@
+interface Props {
+  description: string;
+}
+
+export function PostDescription({ description }: Props) {
+  return (
+    <div className="space-y-4">
+      <p>{description}</p>
+    </div>
+  );
+}
