+++ src/components/post/header.tsx
@@ -1,0 +1,13 @@
+interface Props {
+  title: string;
+  date?: string;
+}
+
+export function PostHeader({ title, date }: Props) {
+  return (
+    <>
+      <h1>{title}</h1>
+      <span>{date}</span>
+    </>
+  );
+}
