+++ src/components/post/tags.tsx
@@ -1,0 +1,9 @@
+interface Props {
+  tags?: string[];
+}
+
+export function PostTags({ tags }: Props) {
+  return tags && (
+    <span>{tags.join(', ')}</span>
+  );
+}
